Fix incorrect next selection after selected item is dragged
For example, take 3 items named 1, 2, and 3. If 2 is dragged so that the
order becomes 2, 1, 3, then the next item being selected was 3 rather
than 1. This was because the `filteredValue` setter was setting the
unfiltered `value` using the `index` property of the item in the array,
but `itemsReordered` had not been handled yet. This meant that the
`index` was still the item's previous position. When the watchCollection
for `filteredItems` was triggered due to the drag causing a re-ordering,
it was setting `filteredValue` to `value`, which had the wrong index.

Fix this by setting `filteredValue` after `itemsReordered` is handled.